Consolidate all test vectors into files with `Web5TestVectors` prefix
